requirements.txt
streamlit==1.32.0
google-generativeai==0.5.0
pdf2image==1.16.3
python-dotenv==1.0.0
Pillow==10.0.0
requests==2.31.0
streamlit-lottie==0.0.5
PyPDF2==3.0.1
python-docx==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.25.2
scikit-learn==1.3.0